feat: infinite scrolling for sessions
