+++ package.json
@@ -280,7 +280,6 @@
 		"cross-spawn": "^7.0.3",
 		"i18next": "^23.8.2",
 		"lodash.isequal": "^4.5.0",
-		"memoize-one": "^6.0.0",
 		"module-alias": "^2.2.3",
 		"moment": "^2.30.1",
 		"tar": "^6.2.0",
