+++ vscode-react/src/types/index.ts
@@ -1,1 +1,2 @@
 export type { Action } from "./contextAction.type";
+export type { AKSessionsTableRowProps } from "./sessions.type";
