+++ vscode-react/src/utilities/index.ts
@@ -1,3 +1,4 @@
 export { vscodeWrapper } from "@react-utilities/vscode";
 export { sendMessage } from "@react-utilities/sendMessage.utils";
 export { getTimePassed } from "@react-utilities/getTimePassed.utils";
+export { getSessionActions } from "@react-utilities/sessions.utils";
