+++ vscode-react/tsconfig.json
@@ -44,6 +44,8 @@
 			"@react-sections/*": ["src/sections/*"],
 			"@react-utilities": ["src/utilities/index"],
 			"@react-utilities/*": ["src/utilities/*"],
+			"@react-types": ["src/types/index"],
+			"@react-types/*": ["src/types/*"],
 			"@sections/*": ["src/sections/*"],
 			"@type": ["../src/types/index"],
 			"@type/*": ["../src/types/*"],
